Release develop 251119

+++ board/zhihe/common/script/fastboot_images.bat
@@ -1,21 +1,27 @@
 :: Script for flashing images using fastboot
 
-::@echo off
+@echo off
 
-echo ###### Start the flashing tool
+fastboot getvar product 2> getvar.tmp
+find "product: a2" "getvar.tmp" >nul
 
-IF EXIST "bootzero-rvbl.bin" (
-    fastboot flash ram bootzero-rvbl.bin || goto :error
-    fastboot reboot
-    fastboot flash ram spl-with-fit-rvbl.bin || goto :error
-    fastboot reboot 
+IF %errorlevel% equ 0 (
+    echo ###### Start flash images ######
 ) ELSE (
-    fastboot flash ram spl-with-fit-rvbl.bin || goto :error
-    fastboot reboot 
+	IF EXIST "bootzero-rvbl.bin" (
+		echo ###### Load flashing tool bootzero & uboot
+		fastboot flash ram bootzero-rvbl.bin || goto :error
+		fastboot reboot
+		fastboot flash ram spl-with-fit-rvbl.bin || goto :error
+		fastboot reboot 
+	) ELSE (
+		echo ###### Load flashing tool uboot
+		fastboot flash ram spl-with-fit-rvbl.bin || goto :error
+		fastboot reboot 
+	)
+	echo ###### Wait for the flashing tool to be ready
+	ping 127.0.0.1 -n 5 >nul
 )
-
-echo ###### Wait for the flashing tool to be ready
-ping 127.0.0.1 -n 5 >nul
 
 echo ###### Flash gpt
 fastboot flash gpt emmc-gpt_primary.img || goto :error
